requirements.txt
click>=8.1.3
requests>=2.28.1
python-dotenv>=0.21.0
rich>=13.3.1
transformers
prompt_toolkit>=4.28.0
